backend/requirements.txt
# Backend Dependencies - FastAPI TODO App
# Core web framework and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and data handling
pymongo==4.9.2
motor==3.6.0
beanie==1.24.0
pydantic>=2.8.0

# Authentication and security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Security and rate limiting
slowapi==0.1.9
redis==5.0.1

# Monitoring and system info
psutil==5.9.6